fix: resource tree gvk resource not found error
